Add action wrapping functions.
Will help generally in using actions, not required but will help.

Mostly will help in making the dew-react component a bit more useful
by actually assiting in the mapping of the objects (e.g. the function 'dispatch'
will be able to take an action creator or creators and return the actions functions.