types again.
full soak has to be the default, partial the helper.

Reason being is no way to set undefined values for keys without it. Can't infer keys of an object on typescript return type (in typings, can in js run time).  Hence only way to say 'returns some fields of X, but I don't know which ones" is to use partial, but ignore undefined instead of !hasKey.

Generally setting undefined values is probably not a good idea anyway